requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
astroid==3.3.4
attrs==24.2.0
black==24.8.0
click==8.1.7
dill==0.3.9
discord==2.3.2
discord.py==2.4.0
flake8==7.1.1
frozenlist==1.4.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.1
python-dotenv==1.0.1
tomlkit==0.13.2
yarl==1.13.1
